* TreeCommand works. * TreeBuilder -stop complaining if the marshaller has parameters   not in the marshalled set.
Committing because we're about to drastically change how TFile is
managed between the event processor and sink.  This is because
Root is not well behaved with two TFile* open on a single file for
write (not that surprising).

We're going to move the file open/close into the event processor
so there's only one TFile object associated with SpecTcl
trees.