SEGFAULT/STACK Smashes on ubuntu bionic:
Pointers to data returned from GetAttribute(XmNuserData) must be large
enough to hold a pointer. Made this clean by including stdint.h
and making the return result uintptr_t - this is used in panes
of the display to hold the index in the pane manager array at which
the pane lives.