Display memory mirroring.
Completed the driver script and ensured that it gets run by the main
C++ program as the interpreter RC file. Also did some didling to deal
with cases where the program would refuse to exit due to the fact
that error does not cause scripts with event loops to exit.