Issue #116 Support an FRIB pipeline filter format.
Got tired of forgetting to run the pre-requisite incorp scripts on
a new clone so modified configure.ac - if an incorp script has not
been run run it using the default tags in

UNIFIED_FMT_VERSION
TCLPLUS_VERSION=
CUTIEPIE_VERSION

At the top of the file.  Stole values from the incorporation scripts
themselves to start with.