Event Sink Pipeline -> rank 1 process
Locate all (I think/hope) instances of the apply command in scripts we
support an change those to applygate  to match the new command. Note
SpecTclRestCommand.tcl's proc named apply is now also named
applygate...but that still leaves a version of that in both rustogramer
and PyQtGui.

I've made issues in both projects...it's also possible in CuteiPie that
SpecTclRestCommand.tcl can be removed.