Issue #127 - mpiSpecTcl segfaults when exiting.
Deal with some interesting issues regarding when atexit has to be
called.  atexit handlers are inherited by forked processes... If Xamine
is the displayer then the fact that SpecTcl forks off a copy of itself
to monitor Xamine means that the exit handler for MPI cleanup must be
established _after_ that point else the bcast to stop the gate pump in
the inherited exithandler by that means it will deadlock.

So atexit was moved after setupDisplay are invoked in operator() rather
than in the init.