Issue #141 - Make rootinterface package work in mpiSpecTcl
Did some dirty work to prevent the Root Event loop from handling
file events  in MPI mode.. why:

In MPI Mode, the event sink rank has no stdin.  If left to itself,
Root's event loop will notice this and exit that process... which,
of course we'd rather not have happen.  Therefore, prior to each
execution of the root event loop we:

1.  Set fNfd - 0 indicating there are not fd signals.
2.  For good measure null out the member that contains the file event
handlers (fFielHandler) and hope to hell this won't result in a memory
leak (or at  least not a bad one).  We don't delete because we don't
want any existing file handlers to decide that the right thing to do is
re-establish themselves (if they try to do so via either their
destructor or a callback).

Yes, this is dirty, I admit it.