Issue #141 - Make rootinterface package work in mpiSpecTcl
Added BeginRun and EndRun API elements to the EventMessage module.
This will either directly trigger the event sink pipeline to
do an OnBegin/OnEnd or invoke the currently stubbed MPISendStateChange
which, when written with an accompanying pump, will cause an event
handler in the Event sink rank to invoke the pipeline OnBegin/OnEnd
methods.