Issue #183 - Popping down the Expand dialog segfaults Xamine.
This is a side change that just ensures there's no race condition
between queuing the display refresh context and stargin the work proc.
There's the chance that future implementations of X11/Xt _might_
run the work proc directoy if "possible" rather than queuing it.
This commit allows for that possibility (not yet true).